:recycle: Update dashbaord component split file
